fix: ProjectDTO moved to the right package
